test-embedding/test2/requirements.txt
python-dotenv 
psycopg2-binary 
requests 
tiktoken
transformers 
torch
paramiko==3.4.0 
sshtunnel==0.4.0 
cryptography==41.0.7
fastapi
google.generativeai
pandas
openpyxl